+++ packages/forms/src/UIForm/fields/Enumeration/screenshot.gif
