+++ packages/forms/src/UIForm/utils/widgets.js
@@ -22,7 +22,7 @@ import Select, { TextModeSelect } from '../fields/Select';
 import Text, { TextTextMode } from '../fields/Text';
 import TextArea, { TextAreaTextMode } from '../fields/TextArea';
 import Toggle, { ToggleTextMode } from '../fields/Toggle';
-import EnumerationWidget from '../fields/EnumerationWidget/EnumerationWidget';
+import EnumerationWidget from '../fields/Enumeration/EnumerationWidget';
 
 const widgets = {
 	// fieldsets
